Changes:
• PSI rating in now an `Int`
• Basic list view added to show PSI/PM2.5 ratings.
• Renamed `ContentView`

+++ Singapore Breathe/API/Models/AirQualityModel.swift
@@ -17,7 +17,7 @@ public struct CombinedRegionMetadatum: Codable {
     public let name: String
     public let location: LabelLocation
     public let pm25Hourly: Int?
-    public let psiHourly: Double?
+    public let psiHourly: Int?
     public let timeStamp: String
     public let updatedTimeStamp: String
 }
